size_t

+++ TwoElectronics/include/Compute_On_GPU.cuh
@@ -6,10 +6,10 @@
 
 
 
-__global__ void pairs_init(particle_pair* pair_array, const long size,
+__global__ void pairs_init(particle_pair* pair_array, const size_t size,
 	const double min_r, const double min_p);
 
-__global__ void  pairs_first_step_on_gpu(particle_pair* first_setp_pair_array, const long size);
+__global__ void  pairs_first_step_on_gpu(particle_pair* first_setp_pair_array, const size_t size);
 
 
 __global__ void pre_second_step_qq(double* QQ_array);
@@ -25,12 +25,12 @@ __global__ void pre_second_step_e1_arr(const double* QQ_array, const double EE0,
 __global__ void pre_second_step_e2_arr(const double* QQ_array, const double EE0, double* E2_array);
 
 __global__ void pairs_second_step_on_gpu
-(particle_pair* second_arr, const long size, double* E1_array, double* E2_array);
+(particle_pair* second_arr, const size_t size, double* E1_array, double* E2_array);
 
 
 __global__ void pairs_second_step_on_gpu_fliter
 (const particle_pair* second_step_pair_array, particle_pair* second_step_pair_array_filter,
-	const long size, unsigned long long* count_z, unsigned long long* count_zz);
+	const size_t size, size_t* count_z, size_t* count_zz);
 
 
 #endif //CALL_CPU_CUH
