+++ TwoElectronics/include/Sci_Constant.h
@@ -82,9 +82,9 @@
 
 
 
-unsigned long long Bytes_Of_Pairs = (Pairs_Total) * sizeof(particle_pair);
-unsigned long long Bytes_Of_Array_Laser = sizeof(double) * 2 * (two_steps);
-const int size_ull = sizeof(unsigned long long);
+size_t Bytes_Of_Pairs = (Pairs_Total) * sizeof(particle_pair);
+size_t Bytes_Of_Array_Laser = sizeof(double) * 2 * (two_steps);
+const int size_ull = sizeof(size_t);
 
 std::string init_file_name = "Initialization.dat";
 std::string first_file_name = "After_First_Step.dat";
